feat(dashboard): add tabbed settings pages to Dispatch shell
Reuse the platform settings tabs (Account, Organization, Team, Billing)
on the Dispatch `/dispatch/settings` routes by extracting a shared
SettingsTabs component and wiring up new DISPATCH_SETTINGS_* routes.
